+++ example/lib/main.dart
@@ -1,24 +1,24 @@
-import 'dart:convert';
 import 'dart:io';
 
 import 'package:firebase_core/firebase_core.dart';
 import 'package:firebase_messaging/firebase_messaging.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_local_notifications/flutter_local_notifications.dart';
 import 'package:quick_chat_wms/quick_chat_wms.dart';
 
-FlutterLocalNotificationsPlugin flutterLocalNotificationsPlugin =
-    FlutterLocalNotificationsPlugin();
+import 'notification_service.dart';
+
+@pragma('vm:entry-point')
+Future<void> _firebaseMessagingBackgroundHandler(RemoteMessage message) async {
+  await Firebase.initializeApp();
+  if (QuickChat.isQuickChatNotification(message.data)) {
+    QuickChat.handleQuickChatBackgroundNotification(message.data);
+  }
+}
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
   await Firebase.initializeApp();
   runApp(const MyApp());
-}
-
-Future<void> _firebaseMessagingBackgroundHandler(RemoteMessage message) async {
-  await Firebase.initializeApp();
-  print('Handling a background message ${message.messageId}');
 }
 
 class MyApp extends StatelessWidget {
@@ -44,21 +44,10 @@ class _MyHomePageState extends State<MyHomePage> {
   @override
   void initState() {
     super.initState();
-    initialize();
-  }
-
-  initialize() async {
     requestPermission();
-    try {
-      await Firebase.initializeApp();
-      FirebaseMessaging.onBackgroundMessage(_firebaseMessagingBackgroundHandler);
-
-    } catch (e) {
-      print(
-          "Quick Chat -------- Firebase not initialized. Make sure to add google-services.json in your app.");
-    }
     getFcmConfigure(context);
-    initLocalNotifications(context);
+    initNotifications(context);
+    QuickChat.initializeNotification(context);
   }
 
   @override
@@ -74,80 +63,41 @@ class _MyHomePageState extends State<MyHomePage> {
               widgetCode: 'YOUR-WIDGET-KEY',
               appBarBackgroundColor: const Color(0XFF0066B3));
         },
-        child: const Center(child: Text('CLICK HERE')),
+        child: Container(
+            //YOUR-CONTAINER
+            ),
       ),
     );
   }
 
-  FlutterLocalNotificationsPlugin flutterLocalNotificationsPlugin =
-      FlutterLocalNotificationsPlugin();
-
   void getFcmConfigure(BuildContext context) async {
+    FirebaseMessaging.instance.getToken().then((token) {
+      QuickChat.setFcmToken(token);
+    });
+
     FirebaseMessaging.onMessage.listen((RemoteMessage message) {
-      debugPrint(
-          "FCM NOTIFICATION ---Foreground message received: ${jsonEncode(message.toMap())}");
-      showLocalNotification(message);
+      if (QuickChat.isQuickChatNotification(message.data)) {
+        QuickChat.showQuickChatNotification(message.data);
+        return;
+      }
     });
 
     FirebaseMessaging.onMessageOpenedApp.listen((RemoteMessage message) {
-      debugPrint(
-          "FCM NOTIFICATION ---User tapped the notification: ${jsonEncode(message.toMap())}");
-      QuickChat.handleNotificationOnClick(context);
+      if (QuickChat.isQuickChatNotification(message.data)) {
+        QuickChat.handleNotificationOnClick(context);
+        return;
+      }
     });
 
     FirebaseMessaging.instance
         .getInitialMessage()
         .then((RemoteMessage? message) {
       if (message != null) {
-        debugPrint(
-            "FCM NOTIFICATION ---App opened from terminated state via notification: ${jsonEncode(message.toMap())}");
-        QuickChat.handleNotificationOnClick(context);
+        if (QuickChat.isQuickChatNotification(message.data)) {
+          QuickChat.handleNotificationOnClick(context);
+          return;
+        }
       }
-    });
-  }
-
-  List<String> messages = [];
-
-  Future<void> showLocalNotification(RemoteMessage message) async {
-    String title = message.notification?.title ?? '';
-    String body = message.notification?.body ?? '';
-    messages.add(body);
-    InboxStyleInformation inboxStyle = InboxStyleInformation(
-      messages,
-      contentTitle: title,
-      summaryText: "Tap to open chat",
-    );
-    var androidDetails = AndroidNotificationDetails(
-      'chat_channel',
-      'Chat Notifications',
-      importance: Importance.high,
-      playSound: true,
-      priority: Priority.high,
-      styleInformation: inboxStyle,
-      onlyAlertOnce: true,
-    );
-
-    var notificationDetails = NotificationDetails(android: androidDetails);
-    debugPrint(
-        "FCM NOTIFICATION ---NOTIFICATION DATA ----------> ${jsonEncode(message.toMap())}");
-
-    await flutterLocalNotificationsPlugin.show(
-      0,
-      title,
-      body,
-      notificationDetails,
-    );
-  }
-
-  void initLocalNotifications(BuildContext context) async {
-    var androidInitialize =
-        const AndroidInitializationSettings('@mipmap/ic_launcher');
-    var initializationSettings =
-        InitializationSettings(android: androidInitialize);
-
-    await flutterLocalNotificationsPlugin.initialize(initializationSettings,
-        onDidReceiveNotificationResponse: (NotificationResponse response) {
-      QuickChat.handleNotificationOnClick(context);
     });
   }
 }